Drop Arduino framework — pure ESP-IDF build
Remove `arduino` from `framework` in platformio.ini. All hardware access
now uses ESP-IDF APIs directly with zero Arduino dependencies.

New files:
- conez_stream.h/cpp: custom Stream/Print base class (ConezStream)
- conez_usb.h/cpp: usb_serial_jtag driver wrapper
- crgb.h/cpp: CRGB/CHSV color types + hsv2rgb (replaces FastLED)
- adc.h/cpp: ESP-IDF ADC1 driver wrapper
- compat.h: constrain/map/random/min/max helpers

Key changes:
- Remove #include <Arduino.h> from all 25+ source files
- Replace Stream/Print inheritance with ConezStream
- Replace FastLED with custom color types (RMT hardware output TBD)
- Replace analogRead/analogReadMilliVolts with ESP-IDF adc1 driver
- Add explicit nvs_flash_init() (was handled by initArduino)
- Remove initArduino() from app_main()
- Fix deprecated APIs (mbedtls, sntp, nvs_entry_find, ets_delay_us)
- Rename lerp() to conez_lerp() to avoid C++20 std::lerp conflict
- Disable CONFIG_MBEDTLS_ROM_MD5 (no MD5 in ESP32-S3 ROM)
- ADC only configures GPIO 1-3 (4-7 are PSRAM SPI, 8-10 are LoRa SPI)
- Update AGENTS.md to reflect pure ESP-IDF architecture

+++ firmware/src/wasm/wasm_imports_string.cpp
@@ -351,7 +351,7 @@ m3ApiRawFunction(m3_str_from_int)
     m3ApiReturnType(int32_t);
     m3ApiGetArg(int32_t, val);
     char buf[16];
-    int n = snprintf(buf, sizeof(buf), "%d", val);
+    int n = snprintf(buf, sizeof(buf), "%ld", (long)val);
     uint32_t dst = pool_alloc(runtime, n + 1);
     if (dst == 0) m3ApiReturn(0);
     uint32_t mem_size = m3_GetMemorySize(runtime);
